Refactor code

+++ src/main/java/com/flowiee/dms/service/storage/impl/DocFieldServiceImpl.java
@@ -2,14 +2,20 @@ package com.flowiee.dms.service.storage.impl;
 
 import com.flowiee.dms.exception.DataInUseException;
 import com.flowiee.dms.entity.storage.DocField;
+import com.flowiee.dms.exception.ResourceNotFoundException;
 import com.flowiee.dms.model.ACTION;
 import com.flowiee.dms.model.MODULE;
 import com.flowiee.dms.repository.storage.DocFieldRepository;
+import com.flowiee.dms.service.BaseService;
 import com.flowiee.dms.service.storage.DocDataService;
 import com.flowiee.dms.service.storage.DocFieldService;
-import com.flowiee.dms.service.system.SystemLogService;
+import com.flowiee.dms.utils.ChangeLog;
 import com.flowiee.dms.utils.MessageUtils;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.flowiee.dms.utils.constants.MasterObject;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,19 +23,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class DocFieldServiceImpl implements DocFieldService {
-    private static final String module = MODULE.STORAGE.name();
-
-    private final DocFieldRepository docFieldRepository;
-    private final DocDataService docDataService;
-    private final SystemLogService systemLogService;
-
-    @Autowired
-    public DocFieldServiceImpl(DocFieldRepository docFieldRepository, DocDataService docDataService, SystemLogService systemLogService) {
-        this.docFieldRepository = docFieldRepository;
-        this.docDataService = docDataService;
-        this.systemLogService = systemLogService;
-    }
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+@RequiredArgsConstructor
+public class DocFieldServiceImpl extends BaseService implements DocFieldService {
+    DocDataService     docDataService;
+    DocFieldRepository docFieldRepository;
 
     @Override
     public List<DocField> findAll() {
@@ -50,16 +48,24 @@ public class DocFieldServiceImpl implements DocFieldService {
     @Override
     public DocField save(DocField docField) {
         DocField docFieldSaved = docFieldRepository.save(docField);
-        systemLogService.writeLog(module, ACTION.STG_DOC_DOCTYPE_CONFIG.name(), "Thêm mới doc_field id=" + docField.getId(), null);
-        logger.info(DocumentInfoServiceImpl.class.getName() + ": Thêm mới doc_field id=" + docField.getId());
+        systemLogService.writeLogCreate(MODULE.STORAGE, ACTION.STG_DOC_DOCTYPE_CONFIG, MasterObject.DocField, "Thêm mới DocField", docFieldSaved.getName());
+        logger.info("{}: Thêm mới doc_field id={}", DocumentInfoServiceImpl.class.getName(), docField.getId());
         return docFieldSaved;
     }
 
     @Override
-    public DocField update(DocField docField, Integer docFieldId) {
-        docField.setId(docFieldId);
-        DocField docFieldUpdated = docFieldRepository.save(docField);
-        systemLogService.writeLog(module, ACTION.STG_DOC_DOCTYPE_CONFIG.name(), "Cập nhật doc_field id=" + docFieldId, null);
+    public DocField update(DocField pDocField, Integer docFieldId) {
+        Optional<DocField> docFieldOpt = this.findById(docFieldId);
+        if (docFieldOpt.isEmpty()) {
+            throw new ResourceNotFoundException("DocField not found!");
+        }
+        DocField docFieldBefore = ObjectUtils.clone(docFieldOpt.get());
+
+        pDocField.setId(docFieldId);
+        DocField docFieldUpdated = docFieldRepository.save(pDocField);
+
+        ChangeLog changeLog = new ChangeLog(docFieldBefore, docFieldUpdated);
+        systemLogService.writeLogUpdate(MODULE.STORAGE, ACTION.STG_DOC_DOCTYPE_CONFIG, MasterObject.DocField, "Cập nhật DocField", changeLog);
         logger.info(DocumentInfoServiceImpl.class.getName() + ": Cập nhật doc_field " + docFieldId);
         return docFieldUpdated;
     }
@@ -67,12 +73,16 @@ public class DocFieldServiceImpl implements DocFieldService {
     @Transactional
     @Override
     public String delete(Integer id) {
+        Optional<DocField> docField = this.findById(id);
+        if (docField.isEmpty()) {
+            throw new ResourceNotFoundException("DocField not found!");
+        }
         if (!docDataService.findByDocField(id).isEmpty()) {
             throw new DataInUseException(MessageUtils.ERROR_DATA_LOCKED);
         }
         docFieldRepository.deleteById(id);
-        systemLogService.writeLog(module, ACTION.STG_DOC_DOCTYPE_CONFIG.name(), "Xóa doc_field id=" + id, null);
-        logger.info(DocumentInfoServiceImpl.class.getName() + ": Xóa doc_field id=" + id);
+        systemLogService.writeLogDelete(MODULE.STORAGE, ACTION.STG_DOC_DOCTYPE_CONFIG, MasterObject.DocField, "Xóa DocField", docField.get().getName());
+        logger.info(DocumentInfoServiceImpl.class.getName() + ": Xóa DocField id=" + id);
         return MessageUtils.DELETE_SUCCESS;
     }
 }
